Fix/bug 68174

+++ apps/common/main/lib/controller/Themes.js
@@ -408,7 +408,7 @@ define([
                         }
                     } else
                     if ( e.key == 'content-theme' ) {
-                        this.setContentTheme(e.originalEvent.newValue, true);
+                        this.setContentTheme(e.originalEvent.newValue, true, false);
                     }
                 }.bind(this))
 
@@ -479,14 +479,15 @@ define([
                 return window.uitheme.iscontentdark;
             },
 
-            setContentTheme: function (mode, force) {
+            setContentTheme: function (mode, force, keep) {
                 var set_dark = mode == 'dark';
                 if ( set_dark != window.uitheme.iscontentdark || force ) {
+                    window.uitheme.iscontentdark = set_dark;
+
                     if ( this.isDarkTheme() )
                         this.api.asc_setContentDarkMode(set_dark);
 
-                    window.uitheme.iscontentdark = mode;
-                    if ( Common.localStorage.getItem('content-theme') != mode )
+                    if ( !(keep === false) && Common.localStorage.getItem('content-theme') != mode )
                         Common.localStorage.setItem('content-theme', mode);
 
                     Common.NotificationCenter.trigger('contenttheme:dark', set_dark);
